Keep the old package structure
