+++ sync/js/src/main/scala/cats/effect/SyncIOCompanionPlatform.scala
@@ -14,14 +14,10 @@
  * limitations under the License.
  */
 
-package cats.effect.sync
+package cats.effect
 
-private object SyncIOConstants {
-  val MaxStackDepth = 512
+import scalajs.js
 
-  val MapK: Byte = 0
-  val FlatMapK: Byte = 1
-  val HandleErrorWithK: Byte = 2
-  val RunTerminusK: Byte = 3
-  val AttemptK: Byte = 4
+private[effect] trait SyncIOCompanionPlatform { this: SyncIO.type =>
+  final def realTimeDate: SyncIO[js.Date] = realTime.map(d => new js.Date(d.toMillis.toDouble))
 }
